feat: Move request logic to base pb service class
